Fix unsafe pointer-to-integer casts with uintptr_t and proper format specifiers
Replace (int) casts on pointers with (uintptr_t) for arithmetic and
(void *) with %p for printf. Fix pointer arithmetic in stack.c pool
allocator, debug trace macros in cthreads.c (18 tr2/tr3/tr4 calls),
waiting_for field type (int -> uintptr_t), memcpy/printf casts in
bootstrap.c, and symbol table writes in ELF/a_out/coff/rose loaders
(int* -> int32_t* for explicit 32-bit nlist format).

Closes #20

+++ osfmk/src/mach_services/lib/libcthreads/cthread_internals.h
@@ -218,7 +218,7 @@ struct cthread {
     struct cthread 	*list;		/* List of all cthreads. */
     mutex_t		cond_mutex;	/* Event the thread is waiting for. */
 #ifdef	WAIT_DEBUG
-	int waiting_for;
+	uintptr_t waiting_for;
 #endif	/* WAIT_DEBUG */
     volatile int	state;		/* The current state (see below). */
     volatile int	flags;		/* The thread's flags (see below). */
